+++ azcopy/jobsList.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"github.com/Azure/azure-storage-azcopy/v10/common"
 	"github.com/Azure/azure-storage-azcopy/v10/jobsAdmin"
+	"sort"
 	"time"
 )
 
@@ -40,7 +41,19 @@ func (c Client) ListJobs(opts ListJobsOptions) (result ListJobsResponse, err err
 		})
 	}
 
+	// before displaying the jobs, sort them accordingly so that they are displayed in a consistent way
+	sortJobs(details)
+
 	return ListJobsResponse{
 		Details: details,
 	}, nil
+}
+
+func sortJobs(jobsDetails []JobDetail) {
+	// sort the jobs so that the latest one is shown first
+	sort.Slice(jobsDetails, func(i, j int) bool {
+		// this function essentially asks whether i should be placed before j
+		// we say yes if the job i is more recent
+		return jobsDetails[i].StartTime.After(jobsDetails[j].StartTime)
+	})
 }
